refactor some imports in host fs
